Funcionamiento de la vista Mi Equipo y mejorado el estilo, tambien añadido algunas mejoras en otras vistas

+++ database/migrations/2025_04_15_154632_create_jugadores_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('imagen_url');
             $table->string('posicion');
             $table->string('equipo_real');
-            $table->decimal('valor', 12);
+            $table->bigInteger('valor');
             $table->integer('puntos')->default(0); // Puedes mantener el campo 'puntos' si prefieres
             $table->timestamps();
         });
